move auth/openid_connect.go in its own package
