+++ auth/oidc/openid_connect.go
@@ -1,4 +1,4 @@
-package auth
+package oidc
 
 import (
 	"context"
